+++ SplunkInteractions/Sources/SplunkInteractions/Destination/OTelDestination.swift
@@ -25,7 +25,7 @@ struct OTelDestination: SplunkInteractionsDestination {
 
     // MARK: - Sending
 
-    func send(actionName: String, elementId: String?, time: Date) {
+    func sendInteraction(actionName: String, elementId: String?, xpath: String?, time: Date) {
 
         let logProvider = OpenTelemetry.instance
             .loggerProvider
@@ -38,8 +38,40 @@ struct OTelDestination: SplunkInteractionsDestination {
         attributes["component"] = .string("ui")
         attributes["action.name"] = .string(actionName)
 
+        if let xpath {
+            attributes["target_xpath"] = .string(xpath)
+        }
+
         if let elementId {
             attributes["target.type"] = .string(elementId)
+        }
+
+        let logRecordBuilder =
+            logProvider
+            .logRecordBuilder()
+            .setTimestamp(time)
+            .setAttributes(attributes)
+
+        // Send event
+        logRecordBuilder.emit()
+    }
+
+    func sendFrustration(xpath: String?, time: Date) {
+
+        let logProvider = OpenTelemetry.instance
+            .loggerProvider
+            .get(
+                instrumentationScopeName: "splunk-interaction"
+            )
+
+        var attributes: [String: AttributeValue] = [:]
+        attributes["event.name"] = .string("frustration")
+        attributes["component"] = .string("user-interaction")
+        attributes["frustration_type"] = .string("rage")
+        attributes["interaction_type"] = .string("tap")
+
+        if let xpath {
+            attributes["target_xpath"] = .string(xpath)
         }
 
         let logRecordBuilder =