+++ SplunkInteractions/Tests/SplunkInteractionsTests/Support/TestInteractionDestination.swift
@@ -20,11 +20,27 @@ import Foundation
 @testable import SplunkInteractions
 
 final class TestInteractionDestination: SplunkInteractionsDestination {
-    var didReceiveInteractionCallCount = 0
-    var actionName: String?
 
-    func send(actionName: String, elementId _: String?, time _: Date) {
+    var didReceiveInteractionCallCount = 0
+    var didReceiveFrustrationCallCount = 0
+    var actionName: String?
+    var lastFrustrationXpath: String?
+
+    func sendInteraction(
+        actionName: String,
+        elementId _: String?,
+        xpath _: String?,
+        time _: Date
+    ) {
         self.actionName = actionName
         didReceiveInteractionCallCount += 1
+    }
+
+    func sendFrustration(
+        xpath: String?,
+        time _: Date
+    ) {
+        lastFrustrationXpath = xpath
+        didReceiveFrustrationCallCount += 1
     }
 }